backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
httpx==0.28.1
pytest==8.3.3
openai==1.50.2
faiss-cpu==1.8.0.post1
numpy==1.26.4
sentence-transformers==3.1.1
pypdf==5.0.1
deep-translator==1.11.4
python-dotenv==1.0.1
